Support Nushell's new engine aka engine-q
Beginning with Nushell v0.60, Nushell's engine
has been completely rewritten, which changes a lot
of Nushell's syntax and adds a lot of new features.
While in development, it was known as "engine-q".

panache-git was originally written with the old
syntax from the old engine (Nushell v0.44 and prior).
It has been updated to accommodate the new syntax
and to take advantage of some of those new features,
such as else-if clauses, records, and modules.