Convert from script to module for Nushell 0.68
In 0.68, the `source` command has been deprecated.
Going forward, the way to utilize custom commands is
to put them into a module and import them with `use`.

This means that `panache-git.nu` had to be converted
from a script to a module so that the main prompt command
could be imported by a user's Nushell environment config.

Also in 0.68, `empty?` has been renamed to `is-empty`.

See https://www.nushell.sh/blog/2022-09-06-nushell-0_68.html