Do not include a prompt indicator so the user can choose their own
Previously, panache-git included a prompt indicator (`> `)
out of the box, and the advice for the user was to remove Nushell's
own configurable prompt indicator by setting it to the empty string.

However, this conflicted with some of Nushell's intended functionality,
such as the ability to use a different prompt indicator in each vi mode.

Now, panache-git does not include a prompt indicator.
The advice is for the user to use the prompt indicator they like,
but to reset the color so it better matches panache-git's colors.

Also, the docs were updated for how Nushell 0.101 handles config,
in which prompt configuration happens in config.nu instead of env.nu